+++ ja2lib/Sector.c
@@ -18,7 +18,7 @@ struct SectorInfo* GetSectorInfoByXY(uint8_t x, uint8_t y) {
   return &SectorInfo[GetSectorID8(x, y)];
 }
 
-BOOLEAN SectorOursAndPeaceful(int16_t sMapX, int16_t sMapY, int8_t bMapZ) {
+BOOLEAN SectorOursAndPeaceful(uint8_t sMapX, uint8_t sMapY, int8_t bMapZ) {
   // if this sector is currently loaded
   if ((sMapX == gWorldSectorX) && (sMapY == gWorldSectorY) && (bMapZ == gbWorldSectorZ)) {
     // and either there are enemies prowling this sector, or combat is in progress
@@ -40,5 +40,5 @@ BOOLEAN SectorOursAndPeaceful(int16_t sMapX, int16_t sMapY, int8_t bMapZ) {
   return (TRUE);
 }
 
-int16_t GetLoadedSectorX() { return gWorldSectorX; }
-int16_t GetLoadedSectorY() { return gWorldSectorY; }
+uint8_t GetLoadedSectorX() { return (uint8_t)gWorldSectorX; }
+uint8_t GetLoadedSectorY() { return (uint8_t)gWorldSectorY; }
